Jan. 3 2:38PM CFigueroa     Initial commit; includes parse_list_contents(), __parse_numeric(),     __parse_float(), __parse_int() and __contains().

+++ Utils/Parser.py
@@ -4,7 +4,7 @@
     within it. Additionally,  the function can  recursively  parse the contents  of an
     element that is also of type List.
 
-    Supported types: list, str, int, float
+    Supported types: list, str, float, int
 """
